Lesson15.Policies. Created PostPolicy, updated the controller and views.

+++ app/Http/Controllers/Admin/PostController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests\Admin\StoreUpdatePostRequest;
 use App\Models\Post;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -17,6 +18,11 @@ class PostController extends Controller
      */
     public function index()
     {
+        $response = Gate::inspect('viewAny', Post::class);
+        if ($response->denied()) {
+            return back()->with('message',$response->message());
+        };
+
         return view('Admin.Pages.post-table', [
             'title' => 'публікації',
             'pageName' => 'Список публікацій',
@@ -31,6 +37,11 @@ class PostController extends Controller
      */
     public function create() 
     {
+        $response = Gate::inspect('create', Post::class);
+        if ($response->denied()) {
+            return back()->with('message', $response->message());
+        };
+
         return view('Admin.Pages.post-form', [
             'title' => 'створити публікацію',
             'pageName' => 'Створити нову публікацію',
@@ -47,6 +58,11 @@ class PostController extends Controller
      */
     public function store(StoreUpdatePostRequest $request)
     {
+        $response = Gate::inspect('create', Post::class);
+        if ($response->denied()) {
+            return back()->with('message',$response->message());
+        };
+
         $post = new Post;
         $post->title = $request->input('title');
         $post->slug = Str::of($request->input('title'))->slug('-');
@@ -55,7 +71,7 @@ class PostController extends Controller
         $post->user_id = $request->user()->user_id;
         $post->save();
         
-        return back()->withSuccess('Публікація була успішно додана.');
+        return back()->withSuccess( __('The post was successfully added.'));
     }
 
     /**
@@ -66,10 +82,17 @@ class PostController extends Controller
      */
     public function show($id)
     {
+        $post = Post::find($id);
+
+        $response = Gate::inspect('view', $post);
+        if ($response->denied()) {
+            return back()->with('message',$response->message());
+        };
+
         return view('Admin.Pages.post-form', [
             'title' => 'переглянути публікацію',
             'pageName' => 'Переглянути публікацію з id=' . $id,
-            'post' => Post::find($id),
+            'post' => $post,
             'method' => 'show',
         ]);
     }
@@ -82,10 +105,17 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+        $post = Post::find($id);
+
+        $response = Gate::inspect('update', $post);
+        if ($response->denied()) {
+            return back()->with('message',$response->message());
+        };
+
         return view('Admin.Pages.post-form', [
             'title' => 'змінити публікацію',
             'pageName' => 'Змінити публікацію з id=' . $id,
-            'post' => Post::find($id),
+            'post' => $post,
             'method' => 'edit',
         ]);
     }
@@ -100,14 +130,19 @@ class PostController extends Controller
     public function update(StoreUpdatePostRequest $request, $id)
     {
         $post = Post::find($id);
+
+        $response = Gate::inspect('update', $post);
+        if ($response->denied()) {
+            return back()->with('message',$response->message());
+        };
+
         $post->title = $request->input('title');
         $post->slug = Str::of($request->input('title'))->slug('-');
         $post->text = $request->input('text');
         $post->likes = $request->input('likes');
         $post->save();
         
-        //return back()->withSuccess('Публікація була успішно змінена.');
-        return redirect()->route('dashboard.posts.edit', ['post' => $id])->withSuccess('Публікація була успішно змінена.');
+        return redirect()->route('dashboard.posts.edit', ['post' => $id])->withSuccess( __('The post was successfully modified.'));
     }
 
     /**
@@ -118,7 +153,15 @@ class PostController extends Controller
      */
     public function destroy($id)
     {
+        $post = Post::find($id);
+
+        $response = Gate::inspect('delete', $post);
+        if ($response->denied()) {
+            return back()->with('message',$response->message());
+        };
+
         Post::destroy($id);
-        return back()->withSuccess('Публікація була успішно видалена.');
+
+        return back()->withSuccess( __('The post was successfully deleted'));
     }
 }
